Allow is_system_tray_available to be called with running event loop Extend Flatpak permissions
